Move external mocks to own directory
