+++ frontend/cypress/integration/create_evaluation_spec.ts
@@ -3,7 +3,7 @@ import { EvaluationSeed } from '../support/evaluation_seed'
 import { evaluationName } from '../support/helpers'
 import NominationPage from '../support/nomination'
 import ProjectPage from '../support/project'
-import { User, users } from '../support/users'
+import { User, users } from '../support/mock/external/users'
 
 const evaluationUserIsFacilitator = () => {
     let seed = new EvaluationSeed({
