+++ frontend/cypress/support/evaluation_seed.ts
@@ -1,5 +1,5 @@
 import { Progression, Question, Role } from '../../src/api/models'
-import { User, getUsers } from './users'
+import { User, getUsers } from './mock/external/users'
 import { evaluationName} from './helpers'
 import { Answer, Action, createAction, Participant, Note, Summary } from './mocks'
 import {
