+++ frontend/cypress/support/mock/external/projects.ts
